separate view project in list and form
