+++ src/router/index.ts
@@ -1,6 +1,6 @@
 import { createRouter, createWebHashHistory, RouteRecordRaw } from "vue-router";
 import Tasks from "../views/Tasks.vue";
-import Projects from "../views/Projects.vue";
+import ProjectList from "../views/Projects/ProjectList.vue";
 
 const routes: RouteRecordRaw[] = [
     {
@@ -11,7 +11,7 @@ const routes: RouteRecordRaw[] = [
     {
         path: '/projects',
         name: 'projects',
-        component: Projects
+        component: ProjectList
     }
 ];
